requirements/base.txt
Django==3.1.3
django-cleanup==4.0.*
django-environ==0.4.5
django-security==0.12.0
Pillow==7.1.*
mysqlclient==2.0.*